api/auth_requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
python-jose[cryptography]
bcrypt
python-multipart
requests
